inventario/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
matplotlib==3.8.3